# API_Document/default_OpenAPI.json
{
  "swagger": "2.0",
  "info": {
    "description": "如意外卖接口文档",
    "version": "1.0",
    "title": "如意外卖"
  },
  "host": "localhost:9000",
  "basePath": "/",
...
  "paths": {
...
    "/dish/{id}": {
      "get": {
        "tags": [
          "菜品管理"
        ],
        "summary": "通过id查询菜品的详细信息",
        "operationId": "getUsingGET_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«Dish对象»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/setmeal": {
      "post": {
        "tags": [
          "套餐管理"
        ],
        "summary": "新增套餐",
        "operationId": "saveUsingPOST_4",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "setmealDto",
            "description": "setmealDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Setmeal对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "套餐管理"
        ],
        "summary": "修改套餐信息",
        "operationId": "updateUsingPUT_4",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "setmealDto",
            "description": "setmealDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Setmeal对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "套餐管理"
        ],
        "summary": "删除套餐",
        "operationId": "deleteUsingDELETE_3",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "ids",
            "required": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/setmeal/dish/{id}": {
      "get": {
        "tags": [
          "套餐管理"
        ],
        "summary": "移动端查看套餐下的菜品信息",
        "operationId": "dishUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«List«Dish对象»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/setmeal/list": {
      "get": {
        "tags": [
          "套餐管理"
        ],
        "summary": "根据条件查询套餐数据",
        "operationId": "listUsingGET_3",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "description": "菜品分类id",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "code",
            "in": "query",
            "description": "编码",
            "required": false,
            "type": "string"
          },
          {
            "name": "createTime",
            "in": "query",
            "description": "创建时间",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "createUser",
            "in": "query",
            "description": "创建人",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "description",
            "in": "query",
            "description": "描述信息",
            "required": false,
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "description": "主键",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "image",
            "in": "query",
            "description": "图片",
            "required": false,
            "type": "string"
          },
          {
            "name": "isDeleted",
            "in": "query",
            "description": "是否删除",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "name",
            "in": "query",
            "description": "套餐名称",
            "required": false,
            "type": "string"
          },
          {
            "name": "price",
            "in": "query",
            "description": "套餐价格",
            "required": false,
            "type": "number"
          },
          {
            "name": "status",
            "in": "query",
            "description": "状态 0:停用 1:启用",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "updateTime",
            "in": "query",
            "description": "更新时间",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "updateUser",
            "in": "query",
            "description": "修改人",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«List«Setmeal对象»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/setmeal/page": {
      "get": {
        "tags": [
          "套餐管理"
        ],
        "summary": "套餐分页查询",
        "operationId": "pageUsingGET_4",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "套餐名称",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页记录数",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«Page»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/setmeal/status/{status}": {
      "post": {
        "tags": [
          "套餐管理"
        ],
        "summary": "修改套餐售卖状态",
        "operationId": "saleUsingPOST_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "ids",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          },
          {
            "name": "status",
            "in": "path",
            "description": "status",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/setmeal/{id}": {
      "get": {
        "tags": [
          "套餐管理"
        ],
        "summary": "通过id获取套餐信息",
        "operationId": "getByIdUsingGET_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«Setmeal对象»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "DishFlavor对象": {
      "type": "object",
      "properties": {
        "createTime": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间",
          "refType": null
        },
        "createUser": {
          "type": "integer",
          "format": "int64",
          "description": "创建人",
          "refType": null
        },
        "dishId": {
          "type": "integer",
          "format": "int64",
          "description": "菜品",
          "refType": null
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "主键",
          "refType": null
        },
        "isDeleted": {
          "type": "integer",
          "format": "int32",
          "description": "是否删除",
          "refType": null
        },
        "name": {
          "type": "string",
          "description": "口味名称",
          "refType": null
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "description": "更新时间",
          "refType": null
        },
        "updateUser": {
          "type": "integer",
          "format": "int64",
          "description": "修改人",
          "refType": null
        },
        "value": {
          "type": "string",
          "description": "口味数据list",
          "refType": null
        }
      },
      "title": "DishFlavor对象",
      "description": "菜品口味关系表"
    },
    "Dish对象": {
      "type": "object",
      "properties": {
        "categoryId": {
          "type": "integer",
          "format": "int64",
          "description": "菜品分类id",
          "refType": null
        },
        "categoryName": {
          "type": "string",
          "refType": null
        },
        "code": {
          "type": "string",
          "description": "商品码",
          "refType": null
        },
        "copies": {
          "type": "integer",
          "format": "int32",
          "refType": null
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间",
          "refType": null
        },
        "createUser": {
          "type": "integer",
          "format": "int64",
          "description": "创建人",
          "refType": null
        },
        "description": {
          "type": "string",
          "description": "描述信息",
          "refType": null
        },
        "flavors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DishFlavor对象"
          },
          "refType": "DishFlavor对象"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "主键",
          "refType": null
        },
        "image": {
          "type": "string",
          "description": "图片",
          "refType": null
        },
        "isDeleted": {
          "type": "integer",
          "format": "int32",
          "description": "是否删除",
          "refType": null
        },
        "name": {
          "type": "string",
          "description": "菜品名称",
          "refType": null
        },
        "price": {
          "type": "number",
          "description": "菜品价格",
          "refType": null
        },
        "sort": {
          "type": "integer",
          "format": "int32",
          "description": "顺序",
          "refType": null
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "description": "0 停售 1 起售",
          "refType": null
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "description": "更新时间",
          "refType": null
        },
        "updateUser": {
          "type": "integer",
          "format": "int64",
          "description": "修改人",
          "refType": null
        }
      },
      "title": "Dish对象",
      "description": "菜品管理"
    },
...
    "OrderItem": {
      "type": "object",
      "properties": {
        "asc": {
          "type": "boolean",
          "refType": null
        },
        "column": {
          "type": "string",
          "refType": null
        }
      },
      "title": "OrderItem"
    },
...
    "Page": {
      "type": "object",
      "properties": {
        "countId": {
          "type": "string",
          "refType": null
        },
        "current": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "maxLimit": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "optimizeCountSql": {
          "type": "boolean",
          "refType": null
        },
        "orders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItem"
          },
          "refType": "OrderItem"
        },
        "pages": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "records": {
          "type": "array",
          "items": {
            "type": "object"
          },
          "refType": "object"
        },
        "searchCount": {
          "type": "boolean",
          "refType": null
        },
        "size": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "total": {
          "type": "integer",
          "format": "int64",
          "refType": null
        }
      },
      "title": "Page"
    },
...
    "R«Dish对象»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "$ref": "#/definitions/Dish对象"
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«Dish对象»"
    },
...
    "R«List«Dish对象»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Dish对象"
          }
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«List«Dish对象»»"
    },
    "R«List«Setmeal对象»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Setmeal对象0"
          }
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«List«Setmeal对象»»"
    },
...
    "R«Page»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "$ref": "#/definitions/Page"
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«Page»"
    },
    "R«Setmeal对象»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "$ref": "#/definitions/Setmeal对象"
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«Setmeal对象»"
    },
...
    "R«string»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "string"
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«string»"
    },
    "SetmealDish对象": {
      "type": "object",
      "properties": {
        "copies": {
          "type": "integer",
          "format": "int32",
          "description": "份数",
          "refType": null
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间",
          "refType": null
        },
        "createUser": {
          "type": "integer",
          "format": "int64",
          "description": "创建人",
          "refType": null
        },
        "dishId": {
          "type": "integer",
          "format": "int64",
          "description": "菜品id",
          "refType": null
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "主键",
          "refType": null
        },
        "isDeleted": {
          "type": "integer",
          "format": "int32",
          "description": "是否删除",
          "refType": null
        },
        "name": {
          "type": "string",
          "description": "菜品名称 （冗余字段）",
          "refType": null
        },
        "price": {
          "type": "number",
          "description": "菜品原价（冗余字段）",
          "refType": null
        },
        "setmealId": {
          "type": "integer",
          "format": "int64",
          "description": "套餐id ",
          "refType": null
        },
        "sort": {
          "type": "integer",
          "format": "int32",
          "description": "排序",
          "refType": null
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "description": "更新时间",
          "refType": null
        },
        "updateUser": {
          "type": "integer",
          "format": "int64",
          "description": "修改人",
          "refType": null
        }
      },
      "title": "SetmealDish对象",
      "description": "套餐菜品关系"
    },
    "Setmeal对象": {
      "type": "object",
      "properties": {
        "categoryId": {
          "type": "integer",
          "format": "int64",
          "description": "菜品分类id",
          "refType": null
        },
        "categoryName": {
          "type": "string",
          "refType": null
        },
        "code": {
          "type": "string",
          "description": "编码",
          "refType": null
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间",
          "refType": null
        },
        "createUser": {
          "type": "integer",
          "format": "int64",
          "description": "创建人",
          "refType": null
        },
        "description": {
          "type": "string",
          "description": "描述信息",
          "refType": null
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "主键",
          "refType": null
        },
        "image": {
          "type": "string",
          "description": "图片",
          "refType": null
        },
        "isDeleted": {
          "type": "integer",
          "format": "int32",
          "description": "是否删除",
          "refType": null
        },
        "name": {
          "type": "string",
          "description": "套餐名称",
          "refType": null
        },
        "price": {
          "type": "number",
          "description": "套餐价格",
          "refType": null
        },
        "setmealDishes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SetmealDish对象"
          },
          "refType": "SetmealDish对象"
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "description": "状态 0:停用 1:启用",
          "refType": null
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "description": "更新时间",
          "refType": null
        },
        "updateUser": {
          "type": "integer",
          "format": "int64",
          "description": "修改人",
          "refType": null
        }
      },
      "title": "Setmeal对象",
      "description": "套餐"
    },
    "Setmeal对象0": {
      "type": "object",
      "properties": {
        "categoryId": {
          "type": "integer",
          "format": "int64",
          "description": "菜品分类id",
          "refType": null
        },
        "code": {
          "type": "string",
          "description": "编码",
          "refType": null
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间",
          "refType": null
        },
        "createUser": {
          "type": "integer",
          "format": "int64",
          "description": "创建人",
          "refType": null
        },
        "description": {
          "type": "string",
          "description": "描述信息",
          "refType": null
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "主键",
          "refType": null
        },
        "image": {
          "type": "string",
          "description": "图片",
          "refType": null
        },
        "isDeleted": {
          "type": "integer",
          "format": "int32",
          "description": "是否删除",
          "refType": null
        },
        "name": {
          "type": "string",
          "description": "套餐名称",
          "refType": null
        },
        "price": {
          "type": "number",
          "description": "套餐价格",
          "refType": null
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "description": "状态 0:停用 1:启用",
          "refType": null
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "description": "更新时间",
          "refType": null
        },
        "updateUser": {
          "type": "integer",
          "format": "int64",
          "description": "修改人",
          "refType": null
        }
      },
      "title": "Setmeal对象0",
      "description": "套餐"
    },
...
  }
}